Update to make it work with python3. This translate was done by Dan Clewley, working to finish it off.

+++ Py6S/Params/geometry.py
@@ -64,9 +64,9 @@ class Geometry:
       
       
       dt = dateutil.parser.parse(datetimestring, dayfirst=True)
-      self.solar_z = 90.0 - Pysolar.GetAltitude(lat, long, dt)
+      self.solar_z = 90.0 - Pysolar.GetAltitude(lat, int, dt)
       
-      az = Pysolar.GetAzimuth(lat, long, dt)
+      az = Pysolar.GetAzimuth(lat, int, dt)
       
       if az < 0:
         self.solar_a = abs(az) + 180
